fix some compile errors and missing headers

+++ src/helics/core/ActionMessage.hpp
@@ -177,7 +177,7 @@ class ActionMessage
     @param[in] buffer_size-- the size of the buffer
     @return the size of the buffer actually used
     */
-    int toByteArray (char *data, size_t buffer_size) const;
+    int toByteArray (char *data, int buffer_size) const;
     /** convert to a string using a reference*/
     void to_string (std::string &data) const;
     /** convert to a byte string*/
@@ -191,11 +191,11 @@ class ActionMessage
     /** convert a command to a byte vector*/
     std::vector<char> to_vector () const;
     /** generate a command from a raw data stream*/
-    int fromByteArray (const char *data, size_t buffer_size);
+    int fromByteArray (const char *data, int buffer_size);
     /** load a command from a packetized stream /ref packetize
     @return the number of bytes used
     */
-    size_t depacketize (const char *data, size_t buffer_size);
+    int depacketize (const char *data, int buffer_size);
     /** read a command from a string*/
     void from_string (const std::string &data);
     /** read a command from a char vector*/
